gcp-data-analytics/requirements.txt
pandas
google-cloud-bigquery
matplotlib
